Rewrite Company class as service ('contao_company.company')
