+++ config/listener.yml
@@ -1,10 +1,10 @@
 services:
-    contao_company.listener.insert_tags:
-        class: Oveleon\ContaoCompanyBundle\EventListener\InsertTagsListener
-        arguments:
-            - '@contao.framework'
-            - '@router.default'
-            - '@request_stack'
-        tags:
-          - { name: contao.hook, hook: replaceInsertTags }
-        public: true
+  contao_company.listener.insert_tags:
+    class: Oveleon\ContaoCompanyBundle\EventListener\InsertTagsListener
+    arguments:
+      - '@contao.framework'
+      - '@router.default'
+      - '@request_stack'
+    tags:
+      - { name: contao.hook, hook: replaceInsertTags }
+    public: true
